agrega test para clubService

+++ src/module/club/service/clubService.js
@@ -3,6 +3,8 @@
  */
 
 const Club = require('../entity/Club')
+const ClubNotDefinedError = require('./error/clubNotDefinedError')
+const ClubIdNotDefinedError = require('./error/clubIdNotDefinedError')
 
 module.exports = class Service {
 
@@ -19,16 +21,28 @@ module.exports = class Service {
     }
 
     getByID(id) {
+        if (id === undefined) {
+            throw new ClubIdNotDefinedError()
+        }
         return this.clubRepository.getByID(id)
     }
 
     saveTeam(team) {
+        if (team === undefined) {
+            throw new ClubNotDefinedError()
+        }
         return this.clubRepository.saveTeam(team)
     }
     deleteTeam(team) {
+        if (team === undefined) {
+            throw new ClubNotDefinedError()
+        }
         return this.clubRepository.deleteTeam(team)
     }
     editTeam(team) {
+        if (team === undefined) {
+            throw new ClubNotDefinedError()
+        }
         return this.clubRepository.editTeam(team)
     }
 }
